split up rust
